app/backend/requirements.txt
Django~=5.0.6
psycopg2-binary>=2.9.9
django-jazzmin
drf-yasg==1.21.5
djangorestframework~=3.15.2
six~=1.16.0
environ~=1.0
minio~=7.2.7
requests~=2.32.3
anyio~=4.4.0
django-admin-list-filter-dropdown~=1.0.3
django-json-widget~=2.0.1
django-minio-storage~=0.5.7
djangorestframework-simplejwt~=5.3.1
exceptiongroup~=1.2.2
h11~=0.14.0
httpcore~=1.0.5
httpx~=0.27.0
idna~=3.7
pillow~=10.4.0
psycopg2~=2.9.9
python-dotenv~=1.0.1
ruamel.yaml~=0.18.6
ruamel.yaml.clib~=0.2.8
sqlparse~=0.5.1

